prepared water reflections for far terrain (disabled due to some unsolved issues that needs to be solved before it looks correct)
water reflections in far terrain hijack the reflection texture for the sea level plane (which has to render far terrain of course to work - this is not working correctly since although far terrain is in its own layer the default layer also renders the far terrain in white where it is transparent)
another issue is that the update of the reflection texture is not instant - which can be seen from Direnni Tower when looking at the sea)

added comman to console - to change current location fast to help debugging the mods

+++ Game/Addons/IncreasedTerrainDistanceMod/Scripts/IncreasedTerrainDistance.cs
@@ -4,10 +4,12 @@
 //Author: Michael Rauter (a.k.a. Nystul)
 //License: MIT License (http://www.opensource.org/licenses/mit-license.php)
 //Version: 1.54
-//Contributors: Lypyl
+//Contributors: Lypyl, Interkarma
 
 // uncomment next line if enhanced sky mod by Lypyl is present
 #define ENHANCED_SKY_AVAILABLE
+
+//#define REFLECTIONSMOD_AVAILABLE
 
 using UnityEngine;
 using System;
@@ -974,11 +976,27 @@ namespace ProjectIncreasedTerrainDistance
 
             setMaterialFogParameters();
 
-            terrainMaterial.SetFloat("_FogFromSkyTex", 0);
+            terrainMaterial.SetInt("_FogFromSkyTex", 0);
 
             //terrainMaterial.SetFloat("_BlendFactor", blendFactor);
             terrainMaterial.SetFloat("_BlendStart", blendStart);
             terrainMaterial.SetFloat("_BlendEnd", blendEnd);
+
+            //#if REFLECTIONSMOD_AVAILABLE
+            //    RenderTexture reflectionSeaTexture = GameObject.Find("ReflectionsMod").GetComponent<ReflectionsMod.UpdateReflectionTextures>().mirrorReflSeaLevel.m_ReflectionTexture;
+            //    if (reflectionSeaTexture != null)
+            //    {
+            //        terrainMaterial.EnableKeyword("ENABLE_WATER_REFLECTIONS");
+            //        terrainMaterial.SetTexture("_SeaReflectionTex", reflectionSeaTexture);
+            //        terrainMaterial.SetInt("_UseSeaReflectionTex", 1);
+            //    }
+            //    else
+            //    {
+            //        terrainMaterial.SetInt("_UseSeaReflectionTex", 0);
+            //    }
+            //#else
+            //    terrainMaterial.SetInt("_UseSeaReflectionTex", 0);
+            //#endif
 
             // Promote material
             terrain.materialTemplate = terrainMaterial;
